+++ eventstore/inmemory/src/test/java/org/occurrent/eventstore/inmemory/InMemoryEventStoreTest.java
@@ -150,7 +150,8 @@ public class InMemoryEventStoreTest {
             // Then
             assertAll(
                     () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
 
@@ -170,7 +171,8 @@ public class InMemoryEventStoreTest {
             // Then
             assertAll(
                     () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(3L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(1L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(3L)
             );
         }
 
@@ -185,7 +187,8 @@ public class InMemoryEventStoreTest {
             // Then
             assertAll(
                     () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(0L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(0L)
             );
         }
         
@@ -204,7 +207,8 @@ public class InMemoryEventStoreTest {
             // Then
             assertAll(
                     () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(2L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
     }