+++ eventstore/mongodb/native/src/test/java/org/occurrent/eventstore/mongodb/nativedriver/MongoEventStoreTest.java
@@ -392,7 +392,8 @@ class MongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
 
@@ -411,7 +412,8 @@ class MongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(3L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(1L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(3L)
             );
         }
 
@@ -423,7 +425,8 @@ class MongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(0L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(0L)
             );
         }
 
@@ -441,7 +444,8 @@ class MongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(2L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
     }