+++ eventstore/mongodb/spring/blocking/src/test/java/org/occurrent/eventstore/mongodb/spring/blocking/SpringMongoEventStoreTest.java
@@ -305,7 +305,8 @@ public class SpringMongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
 
@@ -324,7 +325,8 @@ public class SpringMongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(3L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(1L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(3L)
             );
         }
 
@@ -336,7 +338,8 @@ public class SpringMongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(0L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(0L)
             );
         }
 
@@ -354,7 +357,8 @@ public class SpringMongoEventStoreTest {
             // Then
             assertAll(
                     () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(2L),
+                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
             );
         }
     }